Take into code review - A single constant is defined applicable both to MSRP and HTTP for unknown expiration. - Add constraint ‘NOT NULL’ in filetransfer provider for KEY_FILEICON_EXPIRATION. - For download HTTP file transfer session, file icon information are made immutable. - Define RingingPeriod and SipTransactionTimeout in milliseconds - Rename validity into expiration since it refers to a timestamp (not a duration) - Add a reason code REJECTED_MEDIA_FAILED for file transfers - Miscellanous minor name changes

+++ core/src/com/gsma/rcs/provider/settings/RcsSettings.java
@@ -1025,10 +1025,10 @@ public class RcsSettings {
     /**
      * Get SIP transaction timeout used to wait SIP response
      * 
-     * @return Timeout in seconds
+     * @return Timeout in milliseconds
      */
-    public int getSipTransactionTimeout() {
-        return readInteger(RcsSettingsData.SIP_TRANSACTION_TIMEOUT);
+    public long getSipTransactionTimeout() {
+        return readLong(RcsSettingsData.SIP_TRANSACTION_TIMEOUT);
     }
 
     /**
@@ -1153,10 +1153,10 @@ public class RcsSettings {
     /**
      * Get ringing period
      *
-     * @return Period in seconds
+     * @return Period in milliseconds
      */
-    public int getRingingPeriod() {
-        return readInteger(RcsSettingsData.RINGING_SESSION_PERIOD);
+    public long getRingingPeriod() {
+        return readLong(RcsSettingsData.RINGING_SESSION_PERIOD);
     }
 
     /**
